+++ etc/hatchet-test.sh
@@ -1,6 +1,0 @@
-hatchet install &&
-HATCHET_RETRIES=3 \
-HATCHET_DEPLOY_STRATEGY=git \
-HATCHET_BUILDPACK_BASE="https://github.com/heroku/heroku-buildpack-scala.git" \
-HATCHET_BUILDPACK_BRANCH=$(git name-rev HEAD 2> /dev/null | sed 's#HEAD\ \(.*\)#\1#') \
-rspec spec/