+++ lib/common.sh
@@ -1,17 +1,9 @@
 #!/usr/bin/env bash
 
-export_env_dir() {
-  env_dir=$1
-  whitelist_regex=${2:-''}
-  blacklist_regex=${3:-'^(PATH|GIT_DIR|CPATH|CPPATH|LD_PRELOAD|LIBRARY_PATH|JAVA_OPTS)$'}
-  if [ -d "$env_dir" ]; then
-    for e in $(ls $env_dir); do
-      echo "$e" | grep -E "$whitelist_regex" | grep -qvE "$blacklist_regex" &&
-      export "$e=$(cat $env_dir/$e)"
-      :
-    done
-  fi
-}
+export BUILDPACK_STDLIB_URL="https://lang-common.s3.amazonaws.com/buildpack-stdlib/v7/stdlib.sh"
+
+SBT_0_VERSION_PATTERN='sbt\.version=\(0\.1[1-3]\.[0-9]*\(-[a-zA-Z0-9_]*\)*\)$'
+SBT_1_VERSION_PATTERN='sbt\.version=\(1\.[0-9]*\.[0-9]*\(-[a-zA-Z0-9_]*\)*\)$'
 
 ## SBT 0.10 allows either *.sbt in the root dir, or project/*.scala or .sbt/*.scala
 detect_sbt() {
@@ -62,7 +54,18 @@ _has_buildPropertiesFile() {
 
 _has_playConfig() {
   local ctxDir=$1
-  test -e $ctxDir/conf/application.conf
+  test -e $ctxDir/conf/application.conf ||
+      test "$IS_PLAY_APP" = "true" ||
+      (test -n "$PLAY_CONF_FILE" &&
+          test -e "$PLAY_CONF_FILE" &&
+          test "$IS_PLAY_APP" != "false") ||
+      (# test for default Play 2.3 and 2.4 setup.
+          test -d $ctxDir/project &&
+          test -r $ctxDir/project/plugins.sbt &&
+          test -n "$(grep "addSbtPlugin(\"com.typesafe.play\" % \"sbt-plugin\"" $ctxDir/project/plugins.sbt | grep -v ".*//.*addSbtPlugin")" &&
+          test -r $ctxDir/build.sbt &&
+          test -n "$(grep "enablePlugins(Play" $ctxDir/build.sbt | grep -v ".*//.*enablePlugins(Play")" &&
+          test "$IS_PLAY_APP" != "false")
 }
 
 _has_playPluginsFile() {
@@ -119,9 +122,10 @@ get_supported_play_version() {
 
 get_supported_sbt_version() {
   local ctxDir=$1
+  local sbtVersionPattern=${2:-$SBT_0_VERSION_PATTERN}
   if _has_buildPropertiesFile $ctxDir; then
     sbtVersionLine="$(grep -P '[ \t]*sbt\.version[ \t]*=' "${ctxDir}"/project/build.properties | sed -E -e 's/[ \t\r\n]//g')"
-    sbtVersion=$(expr "$sbtVersionLine" : 'sbt\.version=\(0\.1[1-3]\.[0-9]\+\(-[a-zA-Z0-9_]*\)*\)$')
+    sbtVersion=$(expr "$sbtVersionLine" : "$sbtVersionPattern")
     if [ "$sbtVersion" != 0 ] ; then
       echo "$sbtVersion"
     else
@@ -162,14 +166,14 @@ _download_and_unpack_ivy_cache() {
   local scalaVersion=$2
   local playVersion=$3
 
-  baseUrl="http://lang-jvm.s3.amazonaws.com/sbt/v4/sbt-cache"
+  baseUrl="https://lang-jvm.s3.amazonaws.com/sbt/v8/sbt-cache"
   if [ -n "$playVersion" ]; then
     ivyCacheUrl="$baseUrl-play-${playVersion}_${scalaVersion}.tar.gz"
   else
     ivyCacheUrl="$baseUrl-base.tar.gz"
   fi
 
-  curl --silent --max-time 60 --location $ivyCacheUrl | tar xzm -C $sbtUserHome
+  curl --retry 3 --silent --max-time 60 --location $ivyCacheUrl | tar xzm -C $sbtUserHome
   if [ $? -eq 0 ]; then
     mv $sbtUserHome/.sbt/* $sbtUserHome
     rm -rf $sbtUserHome/.sbt
@@ -181,8 +185,18 @@ _download_and_unpack_ivy_cache() {
 
 has_supported_sbt_version() {
   local ctxDir=$1
-  local supportedVersion="$(get_supported_sbt_version ${ctxDir})"
-  if [ "$supportedVersion" != "" ] ; then
+  local supportedVersion="$(get_supported_sbt_version ${ctxDir} ${SBT_0_VERSION_PATTERN})"
+  if [ -n "$supportedVersion" ] ; then
+    return 0
+  else
+    return 1
+  fi
+}
+
+has_supported_sbt_1_version() {
+  local ctxDir=$1
+  local supportedVersion="$(get_supported_sbt_version ${ctxDir} ${SBT_1_VERSION_PATTERN})"
+  if [ -n "$supportedVersion" ] ; then
     return 0
   else
     return 1
@@ -223,6 +237,10 @@ detect_play_lang() {
   fi
 }
 
+is_app_dir() {
+  test "$1" != "/app"
+}
+
 uses_universal_packaging() {
   local ctxDir=$1
   test -d $ctxDir/target/universal/stage/bin
@@ -245,45 +263,48 @@ universal_packaging_default_web_proc() {
   fi
 }
 
-run_sbt()
-{
+# sed -l basically makes sed replace and buffer through stdin to stdout
+# so you get updates while the command runs and dont wait for the end
+# e.g. sbt stage | indent
+output() {
+  local logfile="$1"
+  local c='s/^/       /'
+
+  case $(uname) in
+      Darwin) tee -a "$logfile" | sed -l "$c";; # mac/bsd sed: -l buffers on line boundaries
+      *)      tee -a "$logfile" | sed -u "$c";; # unix/gnu sed: -u unbuffered (arbitrary) chunks of data
+  esac
+}
+
+install_sbt_extras() {
+  local optDir=${1}
+  local sbtBinDir=${2}
+
+  rm -f ${sbtBinDir}/sbt-launch*.jar #legacy launcher
+  mkdir -p ${sbtBinDir}
+  cp -p ${optDir}/sbt-extras.sh ${sbtBinDir}/sbt-extras
+  cp -p ${optDir}/sbt-wrapper.sh ${sbtBinDir}/sbt
+
+  chmod 0755 ${sbtBinDir}/sbt-extras
+  chmod 0755 ${sbtBinDir}/sbt
+
+  export PATH="${sbtBinDir}:$PATH"
+}
+
+run_sbt() {
   local javaVersion=$1
   local home=$2
   local launcher=$3
   local tasks=$4
+  local buildLogFile=".heroku/sbt-build.log"
 
-  case $(ulimit -u) in
-  32768) # PX Dyno
-    maxSbtHeap=5220
-    ;;
-  *)     # 2X Dyno
-    maxSbtHeap=768
-    ;;
-  esac
+  echo "" > $buildLogFile
 
   status "Running: sbt $tasks"
-  HOME="$home" sbt ${SBT_EXTRAS_OPTS} \
-    -J-Xmx${maxSbtHeap}M \
-    -J-Xms${maxSbtHeap}M \
-    -J-XX:+UseCompressedOops \
-    -sbt-dir $home \
-    -ivy $home/.ivy2 \
-    -sbt-launch-dir $home/launchers \
-    -Duser.home=$home \
-    -Divy.default.ivy.user.dir=$home/.ivy2 \
-    -Dfile.encoding=UTF8 \
-    -Dsbt.global.base=$home \
-    -Dsbt.log.noformat=true \
-    -no-colors -batch \
-    $tasks < /dev/null 2>&1 | indent
+  SBT_HOME="$home" sbt ${tasks} | output $buildLogFile
 
   if [ "${PIPESTATUS[*]}" != "0 0" ]; then
-    error "Failed to run sbt!
-We're sorry this build is failing! If you can't find the issue in application
-code, please send us an email at support@scalingo.com.
-
-Thanks,
-Scalingo"
+    handle_sbt_errors $buildLogFile
   fi
 }
 
@@ -296,4 +317,21 @@ cache_copy() {
     mkdir -p $to_dir/$rel_dir
     cp -pr $from_dir/$rel_dir/. $to_dir/$rel_dir
   fi
+}
+
+install_jdk() {
+  local install_dir=${1}
+
+  let start=$(nowms)
+  JVM_COMMON_BUILDPACK=${JVM_COMMON_BUILDPACK:-https://buildpack-registry.s3.amazonaws.com/buildpacks/heroku/jvm.tgz}
+  mkdir -p /tmp/jvm-common
+  curl --retry 3 --silent --location $JVM_COMMON_BUILDPACK | tar xzm -C /tmp/jvm-common --strip-components=1
+  source /tmp/jvm-common/bin/util
+  source /tmp/jvm-common/bin/java
+  source /tmp/jvm-common/opt/jdbc.sh
+  mtime "jvm-common.install.time" "${start}"
+
+  let start=$(nowms)
+  install_java_with_overlay ${install_dir}
+  mtime "jvm.install.time" "${start}"
 }