+++ spec/sbt_spec.rb
@@ -12,7 +12,7 @@ describe "Sbt" do
       `git commit -am "redeploy" --allow-empty`
       app.push!
       expect(app.output).to include("Running: sbt compile stage")
-      expect(app.output).to include("[info] Done packaging.")
+      expect(app.output).to include("[success]")
     end
   end
 
@@ -31,6 +31,15 @@ describe "Sbt" do
       # expect(app.output).to match("Running: sbt update")
       expect(app.output).not_to match(/Priming Ivy cache/)
       expect(app.output).to match("Running: sbt compile stage")
+    end
+  end
+
+  it "should work with sbt 1.0" do
+    Hatchet::Runner.new("sbt-one-example").deploy do |app|
+      expect(app.output).to include("Running: sbt compile stage")
+      expect(app.output).to include("https://repo.scala-sbt.org/scalasbt/maven-releases/org/scala-sbt/sbt-launch/1.0.0")
+      expect(app.output).not_to include("Main Scala API documentation to")
+      expect(app.output).to include("[info] Done packaging.")
     end
   end
 end